Moved gradualSteps array out of settings Struct, updated motor controller to handle longer run times as input, fixed swipe motions in sleep app

+++ src/components/motor/MotorController.cpp
@@ -18,7 +18,7 @@ void MotorController::Ring(TimerHandle_t xTimer) {
   motorController->RunForDuration(50);
 }
 
-void MotorController::RunForDuration(uint8_t motorDuration) {
+void MotorController::RunForDuration(uint16_t motorDuration) {
   if (motorDuration > 0 && xTimerChangePeriod(shortVib, pdMS_TO_TICKS(motorDuration), 0) == pdPASS && xTimerStart(shortVib, 0) == pdPASS) {
     nrf_gpio_pin_clear(PinMap::Motor);
   }
